Split JVMCloudService into DefaultCloudService
